feat: add requests timeout
