+++ modules/install/config/genvm-manager.yaml
@@ -1,3 +1,6 @@
+# yaml-language-server: $schema=../../../doc/schemas/default-config.json#/manager
+# yaml-language-server: $schema=https://raw.githubusercontent.com/genlayerlabs/genvm/refs/heads/@GENVM_TAG@/doc/schemas/default-config.json#/manager
+
 threads: 4
 blocking_threads: 48
 log_disable: tracing*,polling*,tungstenite*,tokio_tungstenite*
