+++ modules/install/lib/genvm-lua/lib-genvm.lua
@@ -11,8 +11,8 @@ local value2json = require("value2json")
 ---@field log_json fun(val: any): nil
 ---@field sleep_seconds fun(duration: number): nil
 ---@field request
----| fun(ctx, req: { body: nil | string, url: string, headers: table<string, string>, method: string, error_on_status: boolean | nil, json: false | nil, response_body_max_size: integer | nil }): { body: string, status: integer, headers: table<string, string> }
----| fun(ctx, req: { body: nil | string, url: string, headers: table<string, string>, method: string, error_on_status: boolean | nil, json: true, response_body_max_size: integer | nil }): { body: any, status: integer, headers: table<string, string> }
+---| fun(ctx, req: { body: nil | string, url: string, headers: table<string, string>, method: string, error_on_status: boolean | nil, json: false | nil, response_body_max_size: integer | nil, timeout: number | nil }): { body: string, status: integer, headers: table<string, string> }
+---| fun(ctx, req: { body: nil | string, url: string, headers: table<string, string>, method: string, error_on_status: boolean | nil, json: true, response_body_max_size: integer | nil, timeout: number | nil }): { body: any, status: integer, headers: table<string, string> }
 ---@field split_url fun(url: string): nil | { schema: string, port: number | nil, host: string }
 ---@field user_error fun(val: ModuleError): nil
 ---@field base64_decode fun(val: string): string
